@@ -1,35 +1,32 @@
 {
-  "name": "travel",
-  "version": "0.1.0",
+  "name": "travel-vue3",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vue-tsc -b && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^0.19.0",
-    "better-scroll": "^1.15.2",
-    "core-js": "^2.6.5",
-    "fastclick": "^1.0.6",
     "normalize.css": "^8.0.1",
-    "vue": "^2.6.10",
-    "vue-awesome-swiper": "^2.6.7",
-    "vue-router": "^3.0.3",
-    "vuex": "^3.0.1"
+    "pinia": "^3.0.2",
+    "pinia-plugin-persistedstate": "^4.2.0",
+    "vant": "^4.9.19",
+    "vue": "^3.5.13",
+    "vue-router": "4"
   },
   "devDependencies": {
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@vue/cli-plugin-babel": "^3.9.0",
-    "@vue/cli-plugin-eslint": "^3.9.0",
-    "@vue/cli-service": "^3.9.0",
-    "@vue/eslint-config-standard": "^4.0.0",
-    "babel-eslint": "^10.0.1",
-    "babel-plugin-transform-remove-console": "^6.9.4",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "stylus": "^0.54.5",
-    "stylus-loader": "^3.0.2",
-    "vue-template-compiler": "^2.6.10"
-  }
+    "@types/node": "^22.15.14",
+    "@vant/auto-import-resolver": "^1.3.0",
+    "@vitejs/plugin-vue": "^5.2.3",
+    "@vue/tsconfig": "^0.7.0",
+    "sass": "^1.87.0",
+    "typescript": "~5.8.3",
+    "unplugin-auto-import": "^19.2.0",
+    "unplugin-vue-components": "^28.5.0",
+    "vite": "^6.3.5",
+    "vue-tsc": "^2.2.8"
+  },
+  "packageManager": "pnpm@10.8.0"
 }
